+++ searchAreaControl/searchAreaControl.js
@@ -393,33 +393,40 @@
             if (popup && popup.length > 0) {
 
                 // Node
-                popup.find('.sac-ul').not('.sac-ul-top').children('li').find('.sac-node-name').on('click', function () {
-                    var thisNode = $(this);
-                    if (!thisNode.hasClass('sac-node-disabled')) {
-                        if ($that.opt.multiSelect === false) {
-                            popup.find('.sac-node-name').not(this).removeClass('sac-node-selected');
+                popup.find('.sac-ul')
+                    .not('.sac-ul-top')
+                    .children('li')
+                    .find('.sac-node-name')
+                    .on('click', function () {                        
+                        var thisNode = $(this);
+                        if (!thisNode.hasClass('sac-node-disabled')) {
+                            if ($that.opt.multiSelect === false) {
+                                popup.find('.sac-node-name').not(this).removeClass('sac-node-selected');
+                            }
+                            thisNode.toggleClass('sac-node-selected');
+                            $that._applySelection();
                         }
-                        thisNode.toggleClass('sac-node-selected');
-                        $that._applySelection();
-                    }
-                });
+                    });
 
                 // Node parent
                 if (this.opt.multiSelect === true) {
-                    popup.find('.sac-ul.sac-ul-top').children('li').children('.sac-node-name').on('click', function () {
-                        var that = $(this);
-                        var par = that.closest('li');
-                        var parChildren = par.find('.sac-node-name').not(that);
-                        var parSelectedChildren = parChildren.filter('.sac-node-selected');
-                        var parChildrenNum = parChildren.length;
-                        var parSelectedChildrenNum = parSelectedChildren.length;
-                        if (parChildrenNum === parSelectedChildrenNum) {
-                            parChildren.removeClass('sac-node-selected');
-                        } else {
-                            parChildren.addClass('sac-node-selected');
-                        }
-                        $that._applySelection();
-                    });
+                    popup.find('.sac-ul.sac-ul-top')
+                        .children('li')
+                        .children('.sac-node-name')                        
+                        .on('click', function () {                            
+                            var that = $(this);
+                            var par = that.closest('li');
+                            var parChildren = par.find('.sac-node-name').not(that).not('.sac-node-disabled');
+                            var parSelectedChildren = parChildren.filter('.sac-node-selected');
+                            var parChildrenNum = parChildren.length;
+                            var parSelectedChildrenNum = parSelectedChildren.length;
+                            if (parChildrenNum === parSelectedChildrenNum) {
+                                parChildren.removeClass('sac-node-selected');
+                            } else {
+                                parChildren.addClass('sac-node-selected');
+                            }
+                            $that._applySelection();
+                        });
                 }
             }
         },
@@ -552,8 +559,8 @@
         _getSelectedNodes: function () {
             var selectedNodes = [];
             var popup = $('#' + this.popupID);
-            if (popup && popup.length > 0) {
-                var allNodes = popup.find('.sac-ul').not('.sac-ul-top').children('li').find('.sac-node-name');
+            var allNodes = this._getAllNodes();
+            if (popup && popup.length > 0 && allNodes !== null) {                
                 allNodes.filter('.sac-node-selected').each(function () {
                     var node = $(this);
                     var attributes = {};
@@ -566,7 +573,7 @@
                     });
                 });
             }
-            var selectedAll = allNodes.length === selectedNodes.length
+            var selectedAll = (allNodes !== null) ? allNodes.length === selectedNodes.length : false;
             return {
                 selectedAll: selectedAll,
                 selectedNodes: selectedNodes
@@ -629,10 +636,11 @@
          * Select all nodes
          */
         _selectAll: function () {
+            var nodes = this._getAllNodes();
             var popup = $('#' + this.popupID);
-            if (popup && popup.length > 0) {
+            if (popup && popup.length > 0 && nodes !== null) {
                 if (this.opt.multiSelect === true) {
-                    popup.find('.sac-ul').not('.sac-ul-top').children('li').find('.sac-node-name').addClass('sac-node-selected');
+                    nodes.not('.sac-node-disabled').addClass('sac-node-selected');
                 } else {
                     console.warn('Unable to perform diselection due to multiSelect option set to false');
                 }
@@ -643,9 +651,10 @@
          * Diselect all nodes
          */
         _diselectAll: function () {
+            var nodes = this._getAllNodes();
             var popup = $('#' + this.popupID);
-            if (popup && popup.length > 0) {
-                popup.find('.sac-ul').not('.sac-ul-top').children('li').find('.sac-node-name').removeClass('sac-node-selected');                
+            if (popup && popup.length > 0 && nodes !== null) {
+                nodes.removeClass('sac-node-selected');                
             }
         },
 
@@ -653,14 +662,49 @@
          * Invert selection
          */
         _invertSelection: function () {
+            var nodes = this._getAllNodes();
             var popup = $('#' + this.popupID);
             if (popup && popup.length > 0) {
-                if (this.opt.multiSelect === true) {
-                    popup.find('.sac-ul').not('.sac-ul-top').children('li').find('.sac-node-name').toggleClass('sac-node-selected');
+                if (this.opt.multiSelect === true && nodes !== null) {
+                    nodes.not('.sac-node-disabled').toggleClass('sac-node-selected');
                 } else {
                     console.warn('Unable to perform invertion due to multiSelect option set to false');
                 }
             }
+        },
+
+        /**
+         * Enable all nodes
+         */
+        _enableAll: function () {
+            var nodes = this._getAllNodes();
+            var popup = $('#' + this.popupID);
+            if (popup && popup.length > 0 && nodes !== null) {
+                nodes.removeClass('sac-node-disabled');                
+            }
+        },
+
+        /**
+         * Disable all nodes
+         */
+        _disableAll: function () {
+            var nodes = this._getAllNodes();
+            var popup = $('#' + this.popupID);
+            if (popup && popup.length > 0 && nodes !== null) {
+                nodes.addClass('sac-node-disabled');                
+            }
+        },
+
+        /**
+         * Get all nodes
+         */
+        _getAllNodes: function () {
+            var nodes = null;
+            var popup = $('#' + this.popupID);
+            if (popup && popup.length > 0) {
+                nodes = popup.find('.sac-ul').not('.sac-ul-top').children('li').find('.sac-node-name');
+            }
+            return nodes;
         },
 
         /**
@@ -683,22 +727,68 @@
 
         /**
          * Set selected nodes
+         * @param {boolean} allSelected
+         * @param {array} collection
+         * @param {string} byAttribute Optional selection by attribute (If not provided, 'selectionByAttribute' option will be selected)
          */
-        setSelectedNodes: function (allSelected, collection) {
+        setSelectedNodes: function (allSelected, collection, byAttribute) {
             var popup = $('#' + this.popupID);
-            if (popup && popup.length > 0) {
-                var allNodes = popup.find('.sac-ul').not('.sac-ul-top').children('li').find('.sac-node-name');
+            var selectByAttribute = (byAttribute) ? byAttribute : this.opt.selectionByAttribute;
+            var allNodes = this._getAllNodes();
+            if (popup && popup.length > 0 && allNodes !== null) {                
                 if (allSelected === true) {
                     allNodes.addClass('sac-node-selected');
                 } else if (collection && Array.isArray(collection) && collection.length > 0) {
                     // Reset selections
                     this._diselectAll();
                     for (var item in collection) {
-                        allNodes.filter('[' + this.opt.selectionByAttribute + '="' + collection[item] + '"]').addClass('sac-node-selected');
+                        allNodes
+                            .filter('[' + selectByAttribute + '="' + collection[item] + '"]')
+                            .addClass('sac-node-selected');
                     }
                 }
                 this._applySelection();
             }
+        },
+
+        /**
+         * Set disabled nodes
+         * @param {array} collection
+         * @param {boolean} diselectDisabled Diselect nodes to be disabled (if they are selected)
+         * @param {string} byAttribute Optional selection by attribute (If not provided, 'selectionByAttribute' option will be selected)
+         */
+        setDisabledNodes: function (collection, diselectDisabled, byAttribute) {
+            var popup = $('#' + this.popupID);
+            var selectByAttribute = (byAttribute) ? byAttribute : this.opt.selectionByAttribute;
+            var allNodes = this._getAllNodes();
+            if (popup && popup.length > 0 && allNodes !== null) {                
+                if (collection && Array.isArray(collection) && collection.length > 0) {
+                    // Enable all
+                    this._enableAll();
+                    for (var item in collection) {
+                        var nodesToDisable = allNodes.filter('[' + selectByAttribute + '="' + collection[item] + '"]');
+                        if (diselectDisabled === true) {
+                            nodesToDisable.removeClass('sac-node-selected');
+                        }
+                        nodesToDisable.addClass('sac-node-disabled');
+                    }
+                }
+                this._applySelection();
+            }
+        },
+
+        /**
+         * Enable all nodes
+         */
+        enableAllNodes: function () {
+            this._enableAll();
+        },
+
+        /**
+         * Disable all nodes
+         */
+        disableAllNodes: function () {
+            this._disableAll();
         },
 
         /**
@@ -737,6 +827,14 @@
          */
         setDisabled: function (disable) {
             this.$el.prop('disabled', disable);
+        },
+
+        /**
+         * Return popup element if exists
+         */
+        getPopup: function () {
+            var popup = $('#' + this.popupID);
+            return (popup && popup.length > 0) ? popup : null;
         },
 
         /**
@@ -807,6 +905,9 @@
         'clearSelection',
         'setDisabled',
         'updateDatasource',
+        'setDisabledNodes',
+        'enableAllNodes',
+        'disableAllNodes',
         'destroy'
     ];
 
@@ -816,7 +917,8 @@
     $.fn[pluginName].getters = [
         'getSelectedNodes',
         'getDisabled',
-        'getSelectedByAttribute'
+        'getSelectedByAttribute',
+        'getPopup'
     ];
 
     $.fn[pluginName].defaults = {